Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,9 +5,9 @@ char* welcome() {
 }
 
 int main() {
-    int a,b,c;
-    scanf("%d %d %d", &a,&b,&c);
-    float avg=(a+b+c)/3;
-    printf("Average: %.2f", avg);
+    double a,b,c;
+    scanf("%lf %lf %lf", &a,&b,&c);
+    double avg=(a+b+c)/3;
+    printf("Average: %.2lf", avg);
     return 0;
 }
